added an extra solve scenario
also added a feature to test 3 different solve scenarios, taken from games where the solve scenario was available.

each button brings up a different scenario which is auto solved with the animations working correctly.  I wanted to be sure things would work for other scenarios than the original one used.

Added a refresh button for when the auto solve is complete, to refresh the page and to try another scenario