major updates included
undo feature now causes card (or multiple cards) to transition from destination pile back to origin pile in about half a second, and then append to origin pile.

Mainly bug fixes such as cards displaying incorrectly on foundation pile, card transitions not working correctly where card moves too far from destination and then jumps into place.

Fixed issue where an undo which leaves an empty pile failed to render a guide border around the empty pile. Now working.

Changed formatting on foundation piles so that they render differently when empty vs when populated.  When empty a default image is shown, when populated a specific styling is used.

Given that stylying of piles was rendered on a card drop at the end of a drag move, the undo function missed out the stylings required for piles left behind (if empty) and piles repopulated by the undo cards.  This has been coded into the undo function now so pile stylings work.

ONE SIGNIFICANT ISSUE PERSISTS:
There's an issue with the undo function, which sometimes halts after several undos and is unable to read certain elements of the tracking objects related to undo cards; perhaps because the undo information is not correct.  I plan on taking some games all the way to almost completion, then undoing back as far as possible while using the console to find bugs.  It's probably a long painstaking process so I'll handle this one little step at a time.

Other than the undo issue, this iteration is essentially complete with the most major features, automatic solve, and undo, largely working.

The issue with some auto solves not being able to complete seems to have been resolved.  The main issue was that sometimes, the wrapper used to single or multiple cards was not being unrwapped and disposed of, and remained on drop piles and so caused a crash when being examined for 'card' details; the wrapper has no card tracking details, so when the auto solve function is looking for the next suitable card to place on the foundation piles, if the wrapper was covering the next necessary card, the card was never found, and an infinite loop starts until the maximum call stack is reached.

I created a function to ensure wrappers are removed, whether cards are dragged and dropped, dragged but released without dropping, or just clicked and not dragged.  Experimented with many games and, thus far, all auto solves have succeeded.  A future non-coding project related to the above problem would be to prove mathematically, that once all deck cards are face up, the game must have a solutions.  The conclusion is intuitive at the moment, but I'd like to work out a proof which will reassure me that, if an auto solve fails, it must be in the codeing rather than in a card configuration that is impossible to solve.

The next major iteration should include the bug fixes for the undo feature... and later iterations will add features enabling playing of the game on mobile devices where only screen tap can be used (or perhaps drag and drop after tap, if there is such a thing).